requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.4
jinja2==3.1.6
python-multipart==0.0.20
